handled additional info using bloc

+++ lib/src/features/weather/data/models/models.dart
@@ -1,9 +1,9 @@
 class  WeatherModel {
       final double currentTemp;
       final String currentSky;
-      final double currentPressure;
+      final int currentPressure;
       final double currentWindSpeed;
-      final double currentHumidity;
+      final int currentHumidity;
 
   WeatherModel({
     required this.currentTemp, required this.currentSky, required this.currentPressure, required this.currentWindSpeed, required this.currentHumidity});
